:fire: 파워유저 (스케줄링/조회)서비스클래스 통합
