+++ src/main/java/com/codeit/sb01_deokhugam/ranking/poweruser/service/PowerUserService.java
@@ -9,5 +9,7 @@ public interface PowerUserService {
 
 	PageResponse<PowerUserDto> findPowerUsers(GetPowerUsersRequest getPowerUsersRequest);
 
+	void calculateAllPeriodRankings();
+
 	void updateUserCountForPeriod(Period period, Long userNumber);
 }
