--- src/main/java/com/codeit/sb01_deokhugam/ranking/poweruser/service/RankingCalculationService.java
@@ -1,6 +1,0 @@
-package com.codeit.sb01_deokhugam.ranking.poweruser.service;
-
-public interface RankingCalculationService {
-	void calculateAllPeriodRankings();
-
-}
